requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiomysql==0.2.0
azure-storage-blob==12.19.1
azure-storage-file-datalake==12.14.1
prometheus-client==0.20.0
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.2.0
anyio==4.3.0